feat: 45: create new path
These changes make it so that there are 3 additional buttons on the map.
'Start Path' will make it so that any bar marker you click on will
extend a path. 'Stop path' will stop the path building and your path
remains on the screen. 'Clear Path' will clear the path from the screen.